Make seperate file for helper stuff and fix pycodestyly
